src/assistant/requirements.txt
websocket>=0.2.1
websocket-client>=1.8.0
vosk>=0.3.45
colorama>=0.4.6
setuptools>=62.0.0
wheel>=0.38.0
argparse>=1.4.0
pywin32>=306
numpy>=1.20.0
scikit-learn>=0.24.2
pymorphy3>=2.0.4
pymorphy3-dicts-ru>=2.0.4
dotenv>=0.9.9
pyyaml>=6.0
psutil>=7.0.0
pygame>=2.6.1
pyaudio>=0.2.13
pycaw>=20240210
